Convert libnchg to an object library
