+++ CMakeLists.txt
@@ -162,5 +162,3 @@ if(NCHG_ENABLE_TESTING)
   target_compile_definitions(nchg_project_options INTERFACE NCHG_ENABLE_TESTING)
   add_subdirectory(test)
 endif()
-
-include(cmake/Install.cmake)
