--- cmake/nchgConfig.cmake.in
@@ -1,67 +1,0 @@
-# Copyright (C) 2024 Roberto Rossini <roberros@uio.no>
-#
-# SPDX-License-Identifier: GPL-3.0
-#
-# This library is free software: you can redistribute it and/or
-# modify it under the terms of the GNU Public License as published
-# by the Free Software Foundation; either version 3 of the License,
-# or (at your option) any later version.
-#
-# This library is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-# Library General Public License for more details.
-#
-# You should have received a copy of the GNU Public License along
-# with this library.  If not, see
-# <https://www.gnu.org/licenses/>.
-
-@PACKAGE_INIT@
-
-include(CMakeFindDependencyMacro)
-
-find_dependency(
-  Boost
-  CONFIG
-  QUIET
-  REQUIRED)
-
-find_dependency(
-  fmt
-  CONFIG
-  QUIET
-  REQUIRED)
-
-find_dependency(
-  hictk
-  CONFIG
-  QUIET
-  REQUIRED)
-
-find_dependency(
-  HighFive
-  CONFIG
-  QUIET
-  REQUIRED)
-
-find_dependency(
-  phmap
-  CONFIG
-  QUIET
-  REQUIRED)
-
-find_dependency(
-  stocc
-  CONFIG
-  QUIET
-  REQUIRED)
-
-find_dependency(
-  spdlog
-  CONFIG
-  QUIET
-  REQUIRED)
-
-include("${CMAKE_CURRENT_LIST_DIR}/nchgTargets.cmake")
-
-check_required_components(nchg)
